Refactored component loader file into smaller components
