+++ src/bootstrap-btn-sizes.js
@@ -1,0 +1,4 @@
+export default {
+    'lg': 'Large',
+    'sm': 'Small'
+};
