+++ src/bootstrap-contexts.js
@@ -1,0 +1,10 @@
+export default [
+    'primary',
+    'secondary',
+    'success',
+    'info',
+    'warning',
+    'danger',
+    'light',
+    'dark',
+];
